+++ registry_image_save/credentials.py
