+++ registry_image_save/secret_scan_registry.py
